@@ -18,12 +18,12 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.11.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  expansion_tile_card: ^1.1.0
+  expansion_tile_card: ^2.0.0
   intl: ^0.17.0
 
   # The following adds the Cupertino Icons font to your application.
