@@ -1,16 +1,17 @@
 import 'package:flutter/material.dart';
 
-class AppBarGrey extends PreferredSize {
+class AppBarGrey extends StatelessWidget with PreferredSizeWidget {
   static const darkGrey = Color(0xff544C4C); 
   final String textInCenter;
   AppBarGrey(this.textInCenter);
   
   @override
   Size get preferredSize => Size.fromHeight(56);
-
+ 
   @override
   Widget build(BuildContext context) {
-    return AppBar(
+    return PreferredSize( preferredSize: Size.fromHeight(56), child: 
+      AppBar(
         backgroundColor: Colors.white,
         leading: IconButton(
           icon: Icon(
@@ -29,6 +30,7 @@ class AppBarGrey extends PreferredSize {
               color: darkGrey,
             ),
         )
-        );
+      ) 
+    );
   }
 }
